Split files, Define zone.admin_type as Option
