+++ src/admin_type.rs
@@ -1,0 +1,6 @@
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub enum AdminType {
+    NonAdministrative,
+    City,
+    Country,
+}
