+++ src/main.rs
@@ -1,31 +1,30 @@
+extern crate geo;
+extern crate itertools;
 #[macro_use]
 extern crate log;
 extern crate mimir;
 extern crate mimirsbrunn;
+extern crate osmpbfreader;
+#[macro_use]
+extern crate serde_derive;
 extern crate structopt;
 #[macro_use]
 extern crate structopt_derive;
-extern crate osmpbfreader;
-extern crate geo;
-#[macro_use]
-extern crate serde_derive;
-extern crate itertools;
 
 use structopt::StructOpt;
-use mimirsbrunn::osm_reader::admin::{administrative_regions, compute_admin_weight};
 use mimirsbrunn::osm_reader::parse_osm_pbf;
 use mimirsbrunn::osm_reader::OsmPbfReader;
-use mimir::{Coord, Property};
-use std::rc::Rc;
 use itertools::Itertools;
 use mimirsbrunn::boundaries::{build_boundary, make_centroid};
 
+mod zone;
+mod admin_type;
 
 #[derive(StructOpt, Debug)]
 struct Args {
     /// OSM PBF file.
     #[structopt(short = "i", long = "input")]
-    input: String
+    input: String,
 }
 
 fn main() {
@@ -43,14 +42,13 @@ pub fn is_admin(obj: &osmpbfreader::OsmObj) -> bool {
                 .get("boundary")
                 .map_or(false, |v| v == "administrative")
             && 
-            rel.tags
-                .get("admin_level").is_some()
+            rel.tags.get("admin_level").is_some()
         }
         _ => false,
     }
 }
 
-pub fn get_zones(pbf: &mut OsmPbfReader){
+pub fn get_zones(pbf: &mut OsmPbfReader) {
     info!("reading pbf...");
     let objects = pbf.get_objs_and_deps(|o| is_admin(o)).unwrap();
     info!("reading pbf done.");
@@ -65,7 +63,7 @@ pub fn get_zones(pbf: &mut OsmPbfReader){
                 .get("admin_level")
                 .and_then(|s| s.parse().ok());
 
-            // administrative region with name ?
+            // Skip administrative region without name
             let name = match relation.tags.get("name") {
                 Some(val) => val,
                 None => {
@@ -77,7 +75,6 @@ pub fn get_zones(pbf: &mut OsmPbfReader){
                 }
             };
 
-            // admininstrative region without coordinates
             let coord_center = relation
                 .refs
                 .iter()
@@ -96,50 +93,19 @@ pub fn get_zones(pbf: &mut OsmPbfReader){
                 .map(|s| s.to_string())
                 .sorted();
             let boundary = build_boundary(relation, &objects);
-            let zone = Zone {
+            let zone = zone::Zone {
                 id: relation.id.0.to_string(),
                 admin_level: level,
-                admin_type: AdminType::None,
+                admin_type: None,
                 name: name.to_string(),
                 zip_codes: zip_codes,
                 center: coord_center.unwrap_or_else(|| make_centroid(&boundary)),
                 boundary: boundary,
                 parent: None,
-                tags: vec![]
+                tags: vec![],
             };
 
             println!("{:?}", zone);
         }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum AdminType{
-    None,
-    Unknown,
-    City,
-    Country,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Zone {
-    pub id: String,
-    pub admin_level: Option<u32>,
-    pub admin_type: AdminType,
-    pub name: String,
-    pub zip_codes: Vec<String>,
-    pub center: Coord,
-    pub boundary: Option<geo::MultiPolygon<f64>>,
-    pub parent: Option<Rc<Zone>>,
-    pub tags: Vec<Property>,
-    // pub links: Vec<Rc<Zone>>
-}
-
-impl Zone {
-    fn is_admin(&self) -> bool{
-        match(self.admin_type) {
-            AdminType::None => false,
-            _ => true
-        }   
     }
 }
